requirements.txt
python-dotenv==0.19.2
pandas==1.3.5
pyspark==3.1.2
pyarrow==6.0.1
fsspec==2022.1.0
numpy==1.22.1